pyproject.toml
[tool.poetry]
name = "recommendation-systems"
version = "0.1.0"
description = ""
authors = ["Pastório <joaopedropastorio@hotmail.com>"]
readme = "README.md"
packages = [{include = "recommendation_systems"}]

[tool.poetry.dependencies]
python = "^3.10"
ipykernel = "^6.22.0"
pandas = "^2.0.0"
numpy = "^1.24.2"
matplotlib = "^3.7.1"
python-dotenv = "^1.0.0"
plotly = "^5.15.0"
nbformat = "^5.9.0"
fastapi = "^0.100.0"
langchain = "^0.0.235"
openai = "^0.27.8"
uvicorn = "^0.23.1"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
